+++ src/models/Documents/Document.tsx
@@ -1,0 +1,9 @@
+export type Document = {
+    id: number;
+    name: string;
+    type: string;
+    organization: string;
+    description: string;
+    date: string;
+    fileName: string | null;
+}
