+++ src/models/Documents/DocumentOnCreateData.tsx
@@ -1,0 +1,12 @@
+import { Organization } from "../../api/decisionsApi"
+import { MethodicDocumentType } from "./MethodicDocumentType"
+
+export type DocumentOnCreateData = {
+    organizations: Organization[];
+    methodicDocumentTypesItems: MethodicDocumentType[];
+}
+
+
+
+
+
