+++ src/models/Documents/DocumentPost.tsx
@@ -1,0 +1,11 @@
+import { Organization } from "../../api/decisionsApi";
+
+export type DocumentPost = {
+    id: number;
+    name: string;
+    type: number;
+    organization: Organization;
+    description: string;
+    date: string;
+    fileName: string | null;
+}
