+++ src/models/Documents/DocumentWraper.tsx
@@ -1,0 +1,6 @@
+import { DocumentPost } from "./DocumentPost";
+
+export type DocumentWrapper = {
+    MethodicDocument: DocumentPost;
+    fileAsBase64: string | null;
+}
