+++ src/models/Documents/MethodicDocumentType.tsx
@@ -1,0 +1,4 @@
+export type MethodicDocumentType = {
+    text: string;
+    value: string;
+}
